requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Werkzeug==3.0.1
PyJWT==2.8.0
web3==6.15.1
boto3==1.34.34
ipfshttpclient==0.8.0
python-socketio==5.11.1
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
Pillow==10.2.0
requests==2.31.0
python-magic==0.4.27 